extract explanation into file
